// users-service/src/domain/enum/applicationStatus.enum.ts
export enum ApplicationStatus {
  RUN_INITIATED = 'RUN_INITIATED',
  LEAD_CREATED = 'LEAD_CREATED',
  LEAD_VERIFIED = 'LEAD_VERIFIED',
  LEAD_ENHANCED = 'LEAD_ENHANCED',
  CONFIRM_STUDENT_DETAILS = 'CONFIRM_STUDENT_DETAILS',
  APPLY_LOANS = 'APPLY_LOANS',
  SUBMIT_LOANS = 'SUBMIT_LOANS',
  DASHBOARD_DETAILS = 'DASHBOARD_DETAILS',
  EKYC_STATE = 'EKYC_STATE',
}